+++ src/lib/external-option-2-label.ts
@@ -1,0 +1,6 @@
+export function externalOption2label(option, optionsExpParsed) {
+  let context = {};
+  context[optionsExpParsed.item] = option;
+
+  return optionsExpParsed.label(context);
+}