+++ src/lib/external-option-2-value.ts
@@ -1,0 +1,6 @@
+export function externalOption2value(option, optionsExpParsed) {
+  let context = {};
+  context[optionsExpParsed.item] = option;
+
+  return optionsExpParsed.value(context);
+}